@@ -13,6 +13,9 @@
     "react": "^0.14.7",
     "react-dom": "^0.14.7",
     "react-router": "^2.0.0",
+    "bootswatch": "^3.3.7",
+    "firebase": "^3.7.2",
+    "lodash": "^4.17.4",
     "webpack-dev-server": "^1.16.2"
   },
   "devDependencies": {
@@ -20,6 +23,10 @@
     "babel-loader": "^6.2.2",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
-    "webpack": "^1.12.13"
+    "babel-preset-stage-2": "^6.22.0",
+    "css-loader": "^0.27.3",
+    "style-loader": "^0.13.2",
+    "webpack": "^2.2.1"
+
   }
 }
